requirements.txt
aerich==0.8.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
asyncclick==8.1.8.0
asyncpg==0.30.0
atlastk==0.13.3
attrs==25.3.0
boto3==1.37.17
botocore==1.37.17
certifi==2025.1.31
charset-normalizer==3.4.1
ciso8601==2.3.2
click==8.1.8
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
dictdiffer==0.9.0
distro==1.9.0
executing==2.2.0
fastapi==0.115.11
fastapi-jwt-auth==0.5.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
iso8601==2.1.0
jedi==0.19.2
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.49
langchain-openai==0.3.10
langchain-text-splitters==0.3.7
langgraph==0.3.20
langgraph-checkpoint==2.0.23
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.59
langsmith==0.3.19
limits==4.0.1
lxml==5.3.1
marshmallow==3.26.1
matplotlib-inline==0.1.7
multidict==6.2.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.2.4
openai==1.68.2
orjson==3.10.15
ormsgpack==1.9.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
propcache==0.3.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
PyJWT==1.7.1
PyMuPDF==1.25.2
pypdf==5.4.0
pypika-tortoise==0.5.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.11.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SpeechRecognition==3.14.1
SQLAlchemy==2.0.39
stack-data==0.6.3
starlette==0.41.3
tenacity==9.0.0
tiktoken==0.8.0
tornado==6.4.2
tortoise-orm==0.24.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
wcwidth==0.2.13
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
